backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
flask-jwt-extended==4.6.0

